# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",   
    "version": "1.0.2",
    "title": "testdb",
    "termsOfService": "",
    "contact": {
      "email": "shubhamjulaha.rexweb@gmail.com"
    },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "192.168.43.50:3000",
  "basePath": "",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/PerformanceMatrixMainParameters": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Main Parameters"
        ],
        "summary": "This API will provide you the list of all main parameters of PerformanceMatrix",
        "description": "List of all Performance Matrix",
        "operationId": "getPerformance Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Main Parameters"
        ],
        "summary": " This API will add a new main category in performance matrix module ",
        "description": "Add a new performance matrix into db",
        "operationId": "addPerformance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Performance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/PerformanceMatrixSubParameters": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Sub Parameters"
        ],
        "summary": "This API will provide us the list of all sub categories of performnace matrix module",
        "description": "List of all sub Performance Matrix",
        "operationId": "getSubPerformance Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Performance Matrix Sub Parameters"
        ],
        "summary": " This API is for adding new sub category to performance matrix sub parameters ",
        "description": "Add a new sub-performance matrix into db",
        "operationId": "addPerformance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Performance that needs to be added to the store",
            "required": false,
            "schema": {
              "$ref": "#/definitions/subPerformance"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/salesTarget?timeperiod=monthly": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Sales Management Module"
        ],
        "summary": "This API will return monthly target achieved by the employee ",
        "description": "This module will return grouping of user response atabler and training data table on the basis of syllabusId",
        "operationId": "getTarget",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Performance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "percentage": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
    "subPerformance": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "parentId": {
          "type": "integer"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
...
  },
...
}